removed unused CLI flags moved code create_zip_addon.py => build.zip_addon() refactoring
